+++ src/utils.ts
@@ -51,5 +51,10 @@ export function fileNameToUri(gistId: string, filename: string): Uri {
 
 export function getGistLabel(gist: Gist) {
 	const suffix = gist.public ? "" : " (Secret)"
+	return `${Object.keys(gist.files)[0]}${suffix}`
+}
+
+export function getStarredGistLabel(gist: Gist) {
+	const suffix = gist.public ? "" : " (Secret)"
 	return `${gist.owner.login} / ${Object.keys(gist.files)[0]}${suffix}`
 }